Updated to support auto-wiring (Symfony 3.4+)
Symfony 3.4+ makes all services private by default in preparation for
Symfony 4.

This PR auto-wires the Rollout services, which fixes the issue of not
being able to add users to a feature flag in Symfony 3.4.

Also updates the controller to use auto-wired services and the built in
Symfony controller methods, plus type hinting.